Add UIViewController extension and separete files
