+++ BottomHalfModal/SheetContentHeight.swift
@@ -1,0 +1,13 @@
+//
+//  SheetContentHeight.swift
+//  BottomHalfModal
+//
+//  Created by masamichi on 2019/10/09.
+//  Copyright © 2019 merpay, Inc. All rights reserved.
+//
+
+import Foundation
+
+public enum SheetContentHeight {
+    public static let `default`: CGFloat = 320
+}
